requirements.txt
Flask==3.1.2
Flask_SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
reportlab==0.0.1